refactor: 💡 Rename products to catalog
