+++ apps/webstore/src/app/pages/catalog/catalog.component.html
@@ -8,7 +8,7 @@
     </section>
 </div> -->
 
-<div class="product-page-container">
+<div class="catalog-page-container">
     <div class="mini-product-display-container">
         <div class="mini-product"></div>
         <div class="mini-product"></div>
