+++ apps/webstore/src/app/pages/catalog/catalog.component.scss
@@ -1,4 +1,4 @@
-.product-page-container {
+.catalog-page-container {
   display: flex;
   justify-content: center;
   align-items: center;
