+++ apps/webstore/src/app/pages/catalog/catalog.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'webstore-catalog',
+  templateUrl: './catalog.component.html',
+  styleUrls: ['./catalog.component.scss'],
+})
+export class CatalogComponent {
+  constructor() {}
+}
