+++ apps/webstore/src/app/pages/catalog/mini-product-display/mini-product-display.component.spec.ts
@@ -8,9 +8,8 @@ describe('MiniProductDisplayComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ MiniProductDisplayComponent ]
-    })
-    .compileComponents();
+      declarations: [MiniProductDisplayComponent],
+    }).compileComponents();
   });
 
   beforeEach(() => {
