+++ apps/webstore/src/app/pages/catalog/mini-product-display/mini-product-display.component.ts
@@ -3,13 +3,10 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'webstore-mini-product-display',
   templateUrl: './mini-product-display.component.html',
-  styleUrls: ['./mini-product-display.component.scss']
+  styleUrls: ['./mini-product-display.component.scss'],
 })
 export class MiniProductDisplayComponent implements OnInit {
+  constructor() {}
 
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
+  ngOnInit(): void {}
 }
