+++ apps/webstore/src/app/pages/catalog/product-display/product-display.component.ts
@@ -3,13 +3,10 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'webstore-product-display',
   templateUrl: './product-display.component.html',
-  styleUrls: ['./product-display.component.scss']
+  styleUrls: ['./product-display.component.scss'],
 })
 export class ProductDisplayComponent implements OnInit {
+  constructor() {}
 
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
+  ngOnInit(): void {}
 }
