+++ apps/webstore/src/app/pages/catalog/products/quotable/quotable.component.mobile.scss
