--- apps/webstore/src/app/pages/products/products.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'webstore-products',
-  templateUrl: './products.component.html',
-  styleUrls: ['./products.component.scss'],
-})
-export class ProductsComponent {
-  constructor() {}
-}
